pyproject.toml
[tool.poetry]
name = "timetracker"
version = "0.1.0"
description = "Track what windows you spend time on"
authors = ["Violet white <violet.white.dammit@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
svgwrite = "^1.4.1"
SQLAlchemy = "^1.4.25"
CherryPy = "^18.6.1"
xdg = "^5.1.1"
python-libinput = {version="^0.3.0a0",allow-prereleases = true }
pysqlcipher3 = { version = "^1.0.4", optional = true }

[tool.poetry.dev-dependencies]
pytest = "^5.2"
Sphinx = "^4.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
